Change deleteItem mutation to check user permissions

+++ server/src/resolvers/Mutation.js
@@ -40,9 +40,18 @@ const mutations = {
 
     async deleteItem(parent, args, ctx, info) {
         const where = {id: args.id};
-        const item = await ctx.db.query.item({where}, `{id title}`);
+        const item = await ctx.db.query.item({where}, `{id title user {id}}`);
 
-        return ctx.db.mutation.deleteItem({where}, info);
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(
+            permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+
+        if (ownsItem || hasPermissions) {
+            
+            return ctx.db.mutation.deleteItem({where}, info);
+        }
+        throw new Error('You do not have proper permissions to delete taht item!');
     },
 
     async signup(parent, args, ctx, info) {
